[DPE-7482] Remove authentication actions and implement secure auth mechanism (#96)

+++ src/events/provider.py
@@ -160,14 +160,14 @@ class KyuubiClientProviderEvents(BaseEventHandler, WithLogging):
         if not self.charm.unit.is_leader():
             return
 
-        # FIXME: There is not guarantee here
-        auth = AuthenticationManager(cast(DatabaseConnectionInfo, self.context.auth_db))
-        username = f"relation_id_{event.relation.id}"
+        if self.context.auth_db is not None:
+            auth = AuthenticationManager(self.context.auth_db)
+            username = f"relation_id_{event.relation.id}"
 
-        try:
-            auth.delete_user(username)
-        except Exception as e:
-            logger.exception(e)
-            self.charm.unit.status = BlockedStatus(
-                f"Failed to delete user during {KYUUBI_CLIENT_RELATION_NAME} relation broken event"
-            )
+            try:
+                auth.delete_user(username)
+            except Exception as e:
+                logger.exception(e)
+                self.charm.unit.status = BlockedStatus(
+                    f"Failed to delete user during {KYUUBI_CLIENT_RELATION_NAME} relation broken event"
+                )
